==> notebooks/05-ml_training_and_evaluation.ipynb <==
from pyspark.ml.regression import LinearRegression, DecisionTreeRegressor, RandomForestRegressor, GeneralizedLinearRegression
from pyspark.ml.feature import VectorAssembler
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.sql import SparkSession
from pyspark.sql.types import StructType, StringType, DoubleType, IntegerType, DateType
import pyspark.sql.functions as f
import os
from custom_utils import project_base_dir, fuel_type, rolling_window_size, visualisation_dir
import pandas as pd
import matplotlib.pyplot as plt

spark = SparkSession.\
    builder.\
    appName("ml_training-notebook").\
    getOrCreate()

training_schema = StructType() \
      .add("station_uuid",StringType(),True) \
      .add("date",DateType(),True) \
      .add("hour",IntegerType(),True) \
      .add("weekday",IntegerType(),True) \
      .add("deviation",DoubleType(),True) \
      .add("cloudcover",IntegerType(),True) \
      .add("rain",DoubleType(),True) \
      .add("temperature_2m",DoubleType(),True) \
      .add("hour_sin",DoubleType(),True) \
      .add("hour_cos",DoubleType(),True) \
      .add("weekday_sin",DoubleType(),True) \
      .add("weekday_cos",DoubleType(),True)

training_dataframe = spark.read.format("csv") \
      .option("header", True) \
      .schema(training_schema) \
      .load(os.path.join(project_base_dir, "outputs/training_data.csv"))

input_columns = ["cloudcover", "rain", "temperature_2m", "hour_sin", "hour_cos", "weekday_sin", "weekday_cos"]
assembler = VectorAssembler(
    inputCols=input_columns,
    outputCol="features")

data = assembler.transform(training_dataframe)
final_data = data.select("features", "deviation")

train_data, test_data = final_data.randomSplit([0.9, 0.1], seed=42)

def evaluate_predictions(predictions):
    evaluator = RegressionEvaluator(labelCol="deviation", predictionCol="predicted_deviation", metricName="rmse")
    rmse = evaluator.evaluate(predictions)
    print("Root Mean Squared Error (RMSE) on test data: {:.3f}".format(rmse))

    evaluator_r2 = RegressionEvaluator(labelCol="deviation", predictionCol="predicted_deviation", metricName="r2")
    r2 = evaluator_r2.evaluate(predictions)
    print("R-squared (R2) on test data: {:.3f}".format(r2))

def train_and_evaluate_model(model, training_data, test_data):
    l_model = model.fit(training_data)
    predictions = l_model.transform(test_data)
    evaluate_predictions(predictions)
    return l_model

def show_lr_feature_importance(feature_importances, input_columns):
    feature_importance = sorted(list(zip(input_columns, map(abs, feature_importances), feature_importances)), key=lambda x: x[1], reverse=True)

    print("Feature Importance:")
    for feature, importance, coef in feature_importance:
        print("  {}: {:.5f}".format(feature, coef))

def plot_residuals(predictions, model_name):
    residuals_histogram = predictions \
        .withColumn("residuals", f.col("predicted_deviation") - f.col("deviation")) \
        .select('residuals') \
        .rdd.flatMap(lambda x: x) \
        .histogram(50)
    
    ax = pd.DataFrame(
        list(zip(*residuals_histogram)), 
        columns=['bin', 'frequency']
    ).set_index(
        'bin'
    ).plot(
        kind='bar',
        figsize=(10, 5),
        title=f"Distribution of residuals with {model_name}"
    )
    labels = [item.get_text() for item in ax.get_xticklabels()]
    tick_labels = [float(label) for label in labels]
    ax.set_xticklabels([f"{label:.03f}" for label in tick_labels])
    ax.set_xlabel(f"Error between predicted and actual deviation of the {fuel_type} price from its {rolling_window_size}-day rolling average")
    fig = ax.get_figure()
    fig.tight_layout()
    fig.savefig(os.path.join(visualisation_dir, f"{model_name}_prediction_residuals.svg"))
    plt.close(fig)

### Linear regressor

lr = LinearRegression(featuresCol="features", labelCol="deviation", predictionCol="predicted_deviation")

lr_model = train_and_evaluate_model(lr, train_data, test_data)

show_lr_feature_importance(lr_model.coefficients, input_columns)

plot_residuals(lr_model.transform(test_data), LinearRegression.__name__)

### Decision tree regressor

dtr = DecisionTreeRegressor(featuresCol="features", labelCol="deviation", predictionCol="predicted_deviation")

dtr_model = train_and_evaluate_model(dtr, train_data, test_data)

show_lr_feature_importance(dtr_model.featureImportances.toArray(), input_columns)

plot_residuals(dtr_model.transform(test_data), DecisionTreeRegressor.__name__)

### Random forest regressor

rfr = RandomForestRegressor(featuresCol="features", labelCol="deviation", predictionCol="predicted_deviation")

rfr_model = train_and_evaluate_model(rfr, train_data, test_data)

show_lr_feature_importance(rfr_model.featureImportances.toArray(), input_columns)

plot_residuals(rfr_model.transform(test_data), RandomForestRegressor.__name__)

### Generalised linear regression

glm = GeneralizedLinearRegression(featuresCol="features",
                                  labelCol="deviation",
                                  predictionCol="predicted_deviation",
                                  family="gaussian",
                                  link="identity",
                                  maxIter=10,
                                  regParam=0.3)

glm_model = train_and_evaluate_model(glm, train_data, test_data)

show_lr_feature_importance(glm_model.coefficients, input_columns)

plot_residuals(glm_model.transform(test_data), GeneralizedLinearRegression.__name__)
